+++ packages/versioning/src/extensions/cleanup-repo/index.ts
@@ -696,7 +696,6 @@ const extension: VersioningExtension = {
                     } else {
                         const content = [
                             '#!/bin/sh',
-                            '. "$(dirname "$0")/_/husky.sh"',
                             '',
                             cleanupBlock
                         ].join('\n');