+++ packages/versioning/src/extensions/secrets-check/index.ts
@@ -193,7 +193,6 @@ const extension: VersioningExtension = {
                 } else {
                     const content = [
                         '#!/bin/sh',
-                        '. "$(dirname "$0")/_/husky.sh"',
                         '',
                         block
                     ].join('\n');